Updated styling and organized file structure
- Deleted unnecessary css and css.map files
- Created folders for each component
-Implemented the videos.json file
- Updated the styling for all the breakpoints to closely resemble the mockup
- Placed the global css reset on App.scss
