--- src/App.css
@@ -1,39 +1,0 @@
-@font-face {
-  font-family: "Avenir";
-  src: url(../../assets/fonts/AvenirNextLTPro-Regular.otf) format(opentype), url(../../assets/fonts/AvenirNextLTPro.ttf) format(opentype);
-  font-weight: 400;
-}
-@font-face {
-  font-family: "Avenir";
-  src: url(../../assets/fonts/AvenirNextLTPro-Bold.otf) format(opentype), url(../../assets/fonts/AvenirNextLTProBold.ttf) format(opentype);
-  font-weight: 700;
-}
-.container {
-  display: flex;
-  flex-direction: column;
-}
-@media screen and (min-width: 1280px) {
-  .container {
-    flex-direction: row;
-  }
-}
-.container__one {
-  display: flex;
-  flex-direction: column;
-}
-@media screen and (min-width: 1280px) {
-  .container__one {
-    width: 65vw;
-    border-right: 1px solid #e1e1e1;
-  }
-}
-.container__two {
-  display: flex;
-  flex-direction: column;
-}
-@media screen and (min-width: 1280px) {
-  .container__two {
-    margin-left: 1rem;
-    width: 33vw;
-  }
-}/*# sourceMappingURL=App.css.map */
