+++ src/App.scss
@@ -3,6 +3,13 @@
 @use "./styles/partials/variables" as *;
 @use "./styles/partials/mixins" as *;
 
+* {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+  font-family: "Avenir";
+}
+
 .container {
   display: flex;
   flex-direction: column;
