--- src/components/Components/Comment.css
@@ -1,101 +1,0 @@
-@font-face {
-  font-family: "Avenir";
-  src: url(../../assets/fonts/AvenirNextLTPro-Regular.otf) format(opentype), url(../../assets/fonts/AvenirNextLTPro.ttf) format(opentype);
-  font-weight: 400;
-}
-@font-face {
-  font-family: "Avenir";
-  src: url(../../assets/fonts/AvenirNextLTPro-Bold.otf) format(opentype), url(../../assets/fonts/AvenirNextLTProBold.ttf) format(opentype);
-  font-weight: 700;
-}
-.comment {
-  display: flex;
-  flex-direction: column;
-  border-top: 1px solid #e1e1e1;
-}
-@media screen and (min-width: 1280px) {
-  .comment {
-    padding-right: 1rem;
-  }
-}
-.comment:last-child {
-  border-bottom: 1px solid #e1e1e1;
-}
-.comment__top-container {
-  display: flex;
-}
-.comment__photo {
-  margin-top: 1rem;
-  height: 11.25vw;
-  width: 11.25vw;
-  border-radius: 50%;
-  margin-right: 1rem;
-  background-color: #e1e1e1;
-}
-@media screen and (min-width: 768px) {
-  .comment__photo {
-    height: 6.25vw;
-    width: 6.25vw;
-  }
-}
-@media screen and (min-width: 1280px) {
-  .comment__photo {
-    height: 3.75vw;
-    width: 3.75vw;
-  }
-}
-.comment__name {
-  display: flex;
-  align-items: center;
-  font-size: 0.8125rem;
-  line-height: 1.125rem;
-  font-weight: 600;
-}
-@media screen and (min-width: 768px) {
-  .comment__name {
-    font-size: 0.875rem;
-    line-height: 1.375rem;
-  }
-}
-.comment__date {
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-  font-size: 0.75rem;
-  flex-grow: 1;
-  text-align: right;
-  color: #afafaf;
-}
-@media screen and (min-width: 768px) {
-  .comment__date {
-    font-size: 0.875rem;
-    line-height: 1.375rem;
-  }
-}
-.comment__text-container {
-  margin-left: 3.25rem;
-  margin-top: -0.25rem;
-}
-@media screen and (min-width: 768px) {
-  .comment__text-container {
-    margin-left: 4rem;
-    margin-top: -1rem;
-  }
-}
-.comment__text {
-  font-size: 0.8125rem;
-  line-height: 1.125rem;
-  margin-bottom: 1rem;
-}
-@media screen and (min-width: 768px) {
-  .comment__text {
-    font-size: 0.875rem;
-    line-height: 1.375rem;
-  }
-}
-@media screen and (min-width: 1280px) {
-  .comment__text {
-    font-size: 0.875rem;
-    line-height: 1.375rem;
-  }
-}/*# sourceMappingURL=Comment.css.map */
