--- src/components/Components/CommentsForm.css
@@ -1,156 +1,0 @@
-@font-face {
-  font-family: "Avenir";
-  src: url(../../assets/fonts/AvenirNextLTPro-Regular.otf) format(opentype), url(../../assets/fonts/AvenirNextLTPro.ttf) format(opentype);
-  font-weight: 400;
-}
-@font-face {
-  font-family: "Avenir";
-  src: url(../../assets/fonts/AvenirNextLTPro-Bold.otf) format(opentype), url(../../assets/fonts/AvenirNextLTProBold.ttf) format(opentype);
-  font-weight: 700;
-}
-.comments {
-  padding: 1rem;
-}
-@media screen and (min-width: 768px) {
-  .comments {
-    padding: 1.5rem 2.25rem 0rem 2.25rem;
-  }
-}
-@media screen and (min-width: 1280px) {
-  .comments {
-    padding-left: 8.5rem;
-    width: 65vw;
-    padding-right: 1rem;
-  }
-}
-.comments__heading {
-  font-size: 0.8125rem;
-  line-height: 1.125rem;
-  padding-bottom: 1rem;
-}
-@media screen and (min-width: 768px) {
-  .comments__heading {
-    font-size: 0.875rem;
-    line-height: 1.375rem;
-  }
-}
-.comments__container {
-  display: flex;
-}
-@media screen and (min-width: 768px) {
-  .comments__container {
-    padding-bottom: 1.5rem;
-  }
-}
-.comments__avatar {
-  border-radius: 50%;
-  margin-top: 1.5rem;
-  width: 11.25vw;
-  height: 11.25vw;
-  margin-right: 1rem;
-}
-@media screen and (min-width: 768px) {
-  .comments__avatar {
-    width: 5.208vw;
-    height: 5.208vw;
-  }
-}
-@media screen and (min-width: 1280px) {
-  .comments__avatar {
-    height: 3.125vw;
-    width: 3.125vw;
-    margin-right: 2.25rem;
-  }
-}
-.comments__form {
-  display: flex;
-  flex-direction: column;
-  width: 100vw;
-}
-@media screen and (min-width: 768px) {
-  .comments__form {
-    flex-direction: row;
-    align-items: flex-end;
-  }
-}
-@media screen and (min-width: 768px) {
-  .comments__form-container {
-    display: flex;
-    flex-direction: column;
-  }
-}
-.comments__name {
-  padding-bottom: 0.25rem;
-  color: #afafaf;
-  font-size: 0.8125rem;
-  line-height: 1.125rem;
-}
-@media screen and (min-width: 768px) {
-  .comments__name {
-    font-size: 0.75rem;
-    line-height: 1.25rem;
-  }
-}
-.comments__comment-area {
-  height: 31.563vw;
-  resize: none;
-  padding: 0.75rem;
-  border: 1px solid #e1e1e1;
-  border-radius: 0.25rem;
-  width: 73.75vw;
-}
-@media screen and (min-width: 768px) {
-  .comments__comment-area {
-    height: 5.208vw;
-    width: 62vw;
-    margin-right: 1rem;
-  }
-}
-@media screen and (min-width: 1280px) {
-  .comments__comment-area {
-    width: 33vw;
-    height: 3.125vw;
-  }
-}
-.comments__comment-area::-moz-placeholder {
-  color: #000000;
-  font-size: 0.75rem;
-  line-height: 1.25rem;
-}
-.comments__comment-area::placeholder {
-  color: #000000;
-  font-size: 0.75rem;
-  line-height: 1.25rem;
-}
-.comments__comment-area:focus {
-  outline: none;
-  border: 2px solid #0065ad;
-}
-.comments__submit {
-  margin-top: 1rem;
-  height: 12.5vw;
-  border-color: transparent;
-  border-radius: 4px;
-  color: #ffffff;
-  background-color: #0095ff;
-  background-image: url(../../assets/images/add_comment.svg);
-  background-repeat: no-repeat;
-  background-position: center left 0.75rem;
-  background-size: 7vw;
-}
-@media screen and (min-width: 768px) {
-  .comments__submit {
-    height: 5.208vw;
-    width: 19vw;
-    margin-top: 0;
-    font-size: 0.75rem;
-    line-height: 1.25rem;
-    background-size: 2vw;
-  }
-}
-@media screen and (min-width: 1280px) {
-  .comments__submit {
-    height: 3.125vw;
-    width: 13vw;
-  }
-}/*# sourceMappingURL=CommentsForm.css.map */
