--- src/components/Components/CommentsList.css
@@ -1,24 +1,0 @@
-@font-face {
-  font-family: "Avenir";
-  src: url(../../assets/fonts/AvenirNextLTPro-Regular.otf) format(opentype), url(../../assets/fonts/AvenirNextLTPro.ttf) format(opentype);
-  font-weight: 400;
-}
-@font-face {
-  font-family: "Avenir";
-  src: url(../../assets/fonts/AvenirNextLTPro-Bold.otf) format(opentype), url(../../assets/fonts/AvenirNextLTProBold.ttf) format(opentype);
-  font-weight: 700;
-}
-.list {
-  padding: 0rem 1rem 1rem 1rem;
-}
-@media screen and (min-width: 768px) {
-  .list {
-    padding: 0 2.25rem;
-  }
-}
-@media screen and (min-width: 1280px) {
-  .list {
-    padding-left: 8.5rem;
-    padding-right: 0;
-  }
-}/*# sourceMappingURL=CommentsList.css.map */
