--- src/components/Components/Header.css
@@ -1,199 +1,0 @@
-@font-face {
-  font-family: "Avenir";
-  src: url(../../assets/fonts/AvenirNextLTPro-Regular.otf) format(opentype), url(../../assets/fonts/AvenirNextLTPro.ttf) format(opentype);
-  font-weight: 400;
-}
-@font-face {
-  font-family: "Avenir";
-  src: url(../../assets/fonts/AvenirNextLTPro-Bold.otf) format(opentype), url(../../assets/fonts/AvenirNextLTProBold.ttf) format(opentype);
-  font-weight: 700;
-}
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-  font-family: "Avenir";
-}
-
-.header {
-  padding: 1rem;
-}
-@media screen and (min-width: 768px) {
-  .header {
-    display: flex;
-    padding: 1rem 2.25rem;
-    justify-content: space-between;
-    align-items: center;
-  }
-}
-@media screen and (min-width: 1280px) {
-  .header {
-    padding: 1rem 8.5rem;
-  }
-}
-.header__links {
-  display: flex;
-  justify-content: center;
-}
-.header__link {
-  list-style-type: none;
-}
-.header__logo {
-  height: 7.813vw;
-  width: 37.5vw;
-  margin-bottom: 1rem;
-  background-color: #ffffff;
-}
-@media screen and (min-width: 768px) {
-  .header__logo {
-    height: 3.255vw;
-    width: 15.625vw;
-    margin-bottom: 0;
-  }
-}
-@media screen and (min-width: 1280px) {
-  .header__logo {
-    height: 1.953vw;
-    width: 9.375vw;
-    margin-top: 0.5rem;
-  }
-}
-.header__container {
-  display: flex;
-  align-items: center;
-}
-.header__form {
-  display: flex;
-  flex-direction: column;
-}
-@media screen and (min-width: 768px) {
-  .header__form {
-    flex-direction: row;
-    align-items: center;
-  }
-}
-.header__search-container {
-  display: flex;
-  justify-content: space-between;
-}
-.header__search {
-  background-image: url(../../assets/images/search.svg);
-  background-repeat: no-repeat;
-  background-position-y: center;
-  background-position-x: 0.75rem;
-  height: 11.25vw;
-  width: 75vw;
-  max-width: 100vw;
-  padding: 0.75rem;
-  border: 1px solid #e1e1e1;
-  background-color: #fafafa;
-  border-radius: 0.25rem;
-  text-indent: 1.5rem;
-  margin-bottom: 1rem;
-}
-@media screen and (min-width: 768px) {
-  .header__search {
-    width: 31.25vw;
-    height: 5.208vw;
-    margin-bottom: 0;
-    padding: 1rem;
-  }
-}
-@media screen and (min-width: 1280px) {
-  .header__search {
-    width: 18.75vw;
-    height: 3.125vw;
-  }
-}
-.header__search:focus {
-  outline: none;
-  border: 2px solid #0065ad;
-}
-.header__search::-moz-placeholder {
-  color: #000000;
-}
-.header__search::placeholder {
-  color: #000000;
-}
-.header__upload {
-  display: flex;
-  width: 90vw;
-  height: 13.125vw;
-  padding: 0.75rem;
-  border: 1px solid #e1e1e1;
-  border-radius: 0.25rem;
-  background-color: #0095ff;
-  color: #ffffff;
-  align-items: center;
-  text-indent: 25vw;
-  font-size: 1rem;
-  font-weight: 400;
-}
-@media screen and (min-width: 768px) {
-  .header__upload {
-    height: 5.208vw;
-    width: 19.7vw;
-    margin-left: 1rem;
-    text-indent: 0;
-    padding: 1rem 0.25rem;
-    font-size: 0.75rem;
-    line-height: 1.25rem;
-  }
-}
-@media screen and (min-width: 1280px) {
-  .header__upload {
-    height: 3.125vw;
-    width: 11.82vw;
-  }
-}
-.header__upload-img {
-  background-image: url(../../assets/images/upload.svg);
-  background-repeat: no-repeat;
-  height: 7vw;
-  width: 7vw;
-}
-@media screen and (min-width: 768px) {
-  .header__upload-img {
-    height: 3vw;
-    width: 3vw;
-    margin-right: 1rem;
-  }
-}
-@media screen and (min-width: 1280px) {
-  .header__upload-img {
-    height: 2vw;
-    width: 2vw;
-  }
-}
-.header__upload-form {
-  display: none;
-}
-.header__avatar {
-  border-radius: 50%;
-  height: 11.25vw;
-  width: 11.25vw;
-}
-@media screen and (min-width: 768px) {
-  .header__avatar {
-    display: none;
-    height: 5.208vw;
-  }
-}
-.header__avatar-td {
-  display: none;
-}
-@media screen and (min-width: 768px) {
-  .header__avatar-td {
-    display: flex;
-    border-radius: 50%;
-    width: 5.208vw;
-    height: 5.208vw;
-    margin-left: 1rem;
-  }
-}
-@media screen and (min-width: 1280px) {
-  .header__avatar-td {
-    width: 3.125vw;
-    height: 3.125vw;
-  }
-}/*# sourceMappingURL=Header.css.map */
