--- src/components/Components/Videos.css
@@ -1,31 +1,0 @@
-@font-face {
-  font-family: "Avenir";
-  src: url(../../assets/fonts/AvenirNextLTPro-Regular.otf) format(opentype), url(../../assets/fonts/AvenirNextLTPro.ttf) format(opentype);
-  font-weight: 400;
-}
-@font-face {
-  font-family: "Avenir";
-  src: url(../../assets/fonts/AvenirNextLTPro-Bold.otf) format(opentype), url(../../assets/fonts/AvenirNextLTProBold.ttf) format(opentype);
-  font-weight: 700;
-}
-.videos {
-  padding-left: 1rem;
-}
-@media screen and (min-width: 768px) {
-  .videos {
-    padding-left: 2.25rem;
-    padding-top: 1.5rem;
-  }
-}
-@media screen and (min-width: 1280px) {
-  .videos {
-    padding-left: 0;
-  }
-}
-.videos__heading {
-  font-size: 0.75rem;
-  line-height: 1.25rem;
-  color: #afafaf;
-  padding-bottom: 1rem;
-  font-weight: 400;
-}/*# sourceMappingURL=Videos.css.map */
