+++ src/components/Videos/Videos.scss
@@ -5,6 +5,7 @@
 
 .videos {
   padding-left: 1rem;
+  padding-right: 1rem;
 
   @include tablet {
     padding-left: 2.25rem;
